remove info alert from unit data page (#4)
On the page https://heroscapebuilder.com/data the UI/UX is poor. It is not obvious that you can filter and sort columns by clicking the options menu for each column. This page is extremely useful when trying to find units that meet specific criteria. 

Can you please put some effort into improving the overall UI and UX of this page to make it easier when searching and filtering units?